Make OTel resource attributes signal independent
(cherry picked from commit 6bc5d93c4833c32720708db8cd5898dc120b2c5a)
